Avatar scale type fixed to double

+++ Donde.Augmentor/ViewModels/AvatarConfigurationViewModel.cs
@@ -17,9 +17,9 @@
 
     public class AvatarScale
     {
-        public int ScaleX { get; set; }
-        public int ScaleY { get; set; }
-        public int ScaleZ { get; set; }
+        public double ScaleX { get; set; }
+        public double ScaleY { get; set; }
+        public double ScaleZ { get; set; }
     }
 
     public class AvatarAnimation
